esame-backend/requirements.txt
Flask==2.3.2
flask_cors==4.0.0
mysql_connector_repackaged==0.3.1
pandas==2.0.3
